+++ packages/dnr-rulesets/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0-alpha.0] - 2025-04-30
+
+### Changed
+
+- Updated `@adguard/tsurlfilter` to `v3.3.0-alpha.0`.
+
+### Removed
+
+- Byte ranges from metadata ruleset.
+
+[3.0.0-alpha.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/dnr-rulesets-v3.0.0-alpha.0
+
 ## [2.0.0] - 2025-02-28
 
 ## Added