+++ packages/tsurlfilter/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0-alpha.0] - 2025-04-30
+
+### Removed
+
+- Byte ranges management for rulesets.
+
+[3.3.0-alpha.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.3.0-alpha.0
+
 ## [3.2.3] - 2025-04-15
 
 ### Fixed