+++ packages/tswebextension/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0-alpha.0] - 2025-04-30
+
+### Changed
+
+- Updated `@adguard/tsurlfilter` to `v3.3.0-alpha.0`.
+- Instead of byte ranges, rulesets will be synchronized to IndexedDB internally.
+
+### Removed
+
+- `FiltersApi` export.
+
+[3.1.0-alpha.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v3.1.0-alpha.0
+
 ## [3.0.2] - 2025-04-15
 
 ### Changed